Revert "ACRDDF-73 - Creates the template following the paths on the swagger file."
This reverts commit 2ee8aa8bad65f51cc02d951281263b92dc1e0e70.
