+++ .gitignore
@@ -32,3 +32,6 @@ build/
 
 ### VS Code ###
 .vscode/
+
+### Generate-it ###
+.openapi-nodegen
