--- .openapi-nodegen/helpers/swaggerToSpringPath.js
@@ -1,8 +1,0 @@
-const regex = /:(\w+)(?=\/|$)/gm;
-
-/**
- * Transforms the pattern ':param' into '{param}'
- * @param str the path, for example '/vehicles/:vin/:cbs'
- * @returns {string}
- */
-exports.default = (str = '') => str.replace(regex, '{$1}');
