--- .openapi-nodegen/helpers/toDto.js
@@ -1,2 +1,0 @@
-exports.default = (name) =>
-    ! name || name.includes('.') ? name : `com.acrontum.template.dtos.${name}Dto`;
