--- .openapi-nodegen/helpers/trim.js
@@ -1,7 +1,0 @@
-/**
- * Cleans all spaces before and after the text
- * @param str
- * @returns {string}
- */
-exports.default = (str = '') =>
-    str.replace(/^\s+|\s+$/g, '');
