+++ docker-compose.yml
@@ -14,7 +14,7 @@ services:
       - SPRING_DATASOURCE_USERNAME=postgres
       - SPRING_DATASOURCE_PASSWORD=postgres_password
   db:
-    image: 'postgres:14-alpine'
+    image: 'postgres:13.1-alpine'
     container_name: db
     environment:
       - POSTGRES_DB=app_db
