+++ src/main/java/com/acrontum/template/mappers/___op.java.njk
@@ -1,0 +1,42 @@
+{%-set camel_case_operation = _.camelCase(_.deburr(operation_name)).replace(' ','')-%}
+{%-set pascal_case_operation = ucFirst(camel_case_operation)-%}
+package com.acrontum.template.mappers;
+
+import com.acrontum.template.dtos.{{ pascal_case_operation }}Dto;
+import com.acrontum.template.models.{{ pascal_case_operation }}Model;
+import java.util.List;
+import org.mapstruct.Mapper;
+import org.mapstruct.ReportingPolicy;
+
+/**
+ * The interface {{ pascal_case_operation }} mapper.
+ */
+@Mapper(
+        componentModel = "spring",
+        unmappedTargetPolicy = ReportingPolicy.IGNORE
+)
+public interface {{ pascal_case_operation }}Mapper {
+    /**
+     * To dto {{ camel_case_operation }} dto.
+     *
+     * @param {{ camel_case_operation }}Model the {{ camel_case_operation }}
+     * @return the {{ camel_case_operation }} dto
+     */
+    {{ pascal_case_operation }}Dto toDto({{ pascal_case_operation }}Model {{ camel_case_operation }}Model);
+
+    /**
+     * To dto list.
+     *
+     * @param {{ camel_case_operation }}Models the {{ camel_case_operation }}s
+     * @return the list
+     */
+    List<{{ pascal_case_operation }}Dto> toDto(List<{{ pascal_case_operation }}Model> {{ camel_case_operation }}Models);
+
+    /**
+     * To entity {{ camel_case_operation }}.
+     *
+     * @param {{ camel_case_operation }}Dto the {{ camel_case_operation }} dto
+     * @return the {{ camel_case_operation }}
+     */
+    {{ pascal_case_operation }}Model toEntity({{ pascal_case_operation }}Dto {{ camel_case_operation }}Dto);
+}
