+++ src/main/java/com/acrontum/template/repositories/___op.java.njk
@@ -2,7 +2,7 @@
 {%-set pascal_case_operation = ucFirst(camel_case_operation)-%}
 package com.acrontum.template.repositories;
 
-import com.acrontum.template.models.{{ pascal_case_operation }};
+import com.acrontum.template.models.{{ pascal_case_operation }}Model;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -11,12 +11,12 @@ import org.springframework.stereotype.Repository;
  * The interface {{ pascal_case_operation }} repository.
  */
 @Repository
-public interface {{ pascal_case_operation }}Repository extends JpaRepository<{{ pascal_case_operation }}, Integer> {
+public interface {{ pascal_case_operation }}Repository extends JpaRepository<{{ pascal_case_operation }}Model, Integer> {
 
     /**
      * Find by deleted at is null list.
      *
      * @return the list
      */
-    List<{{ pascal_case_operation }}> findByDeletedAtIsNull();
+    List<{{ pascal_case_operation }}Model> findByDeletedAtIsNull();
 }
